% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_headr.R
\name{fun_headr}
\alias{fun_headr}
\title{Create a (fun) R Script Header}
\usage{
fun_headr(name)
}
\arguments{
\item{fun_headr}{}
}
\description{
The fun_headr() function is similar to the hot keys
'cmd + shift + R' for creating a section header.
}
\examples{
# use this to create a fun section header in any .R file
# (with RStudio open)
# fun_headr("packages")

}
