% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodenuff_collab.R
\name{goodenuff_collab}
\alias{goodenuff_collab}
\title{goodenuff_collab}
\usage{
goodenuff_collab()
}
\description{
Create collaboration files: CITATION, requirements.txt, and LICENSE (source: https://creativecommons.org/publicdomain/zero/1.0/legalcode.txt)
}
