% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodenuff_header.R
\name{goodenuff_header}
\alias{goodenuff_header}
\title{Insert goodenuff script header}
\usage{
goodenuff_header()
}
\description{
The goodenuff_header() function can be used in a new R script
    to create a header. The header includes sections for:
- what the code creates
- who the author is
- MIT License
- Version
}
\examples{
# use this to create a script header (with RStudio open)
# goodenuff_header()

}
