% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodenuff_project.R
\name{goodenuff_project}
\alias{goodenuff_project}
\title{goodenuff_project}
\usage{
goodenuff_project(project_name, parent_folder)
}
\arguments{
\item{project_name}{name of your project}

\item{parent_folder}{the parent folder of your project}
}
\description{
Create and open a new 'good enough' project.
}
