% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rheadr.R
\name{rheadr}
\alias{rheadr}
\title{rheadr}
\usage{
rheadr(name)
}
\arguments{
\item{name}{}
}
\description{
The rheadr function is similar to the hot keys 'cmd + shift + R'
for creating a section header.
}
\examples{
# use this to create a section header in a .R file
rheadr("packages")

}
