% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topr.R
\name{topr}
\alias{topr}
\title{topr}
\usage{
topr()
}
\description{
The topr function can be used in a new R script to create a
header. The header includes sections for:
- what the code creates
- who the author is
- MIT License
- Version
}
\examples{
# use this to create a script header

}
